implemented interface pattern.
