--- src/main/java/org/smartwork/json/JsonReader.java
@@ -1,3 +1,0 @@
-package org.smartwork.json;
-
-public interface JsonReader {}
