+++ src/main/java/org/smartwork/json/intf/JsonReader.java
@@ -1,0 +1,9 @@
+package org.smartwork.json.intf;
+
+public interface JsonReader {
+
+	public void execute();
+	
+	public String getJson(String path);
+	
+}
