+++ src/main/resources/resources.properties
@@ -1,3 +1,3 @@
-json.response.by.url=http://demo8189162.mockable.io/getJSONResponse
+json.response.by.url=http://demo3130008.mockable.io/getJSONResponse
 json.response.by.file=sample.json
 json.url.timeout=1000
